fix: correct tool binding and correct pattern pattern now follows default viewer_setup binding
