+++ src/viewer.ts
@@ -82,7 +82,6 @@ import {
   observeWatchable,
   TrackableValue,
 } from "#src/trackable_value.js";
-import { CommandPalette } from "#src/ui/command_palette.js";
 import {
   LayerArchiveCountWidget,
   LayerListPanel,
@@ -108,7 +107,7 @@ import {
 import { AutomaticallyFocusedElement } from "#src/util/automatic_focus.js";
 import { TrackableRGB } from "#src/util/color.js";
 import type { Borrowed, Owned } from "#src/util/disposable.js";
-import { RefCounted, registerEventListener } from "#src/util/disposable.js";
+import { RefCounted } from "#src/util/disposable.js";
 import { removeFromParent } from "#src/util/dom.js";
 import type { ActionEvent } from "#src/util/event_action_map.js";
 import { registerActionListener } from "#src/util/event_action_map.js";
@@ -1140,33 +1139,6 @@ export class Viewer extends RefCounted implements ViewerState {
     );
     this.bindAction("toggle-show-statistics", () => this.showStatistics());
 
-    // Guard prevents double-open when both the element-level action listener and
-    // the document capture listener fire for the same keypress.
-    let openPalette: CommandPalette | undefined;
-    const openCommandPalette = () => {
-      if (openPalette !== undefined && !openPalette.wasDisposed) return;
-      const prevFocused = document.activeElement;
-      const dispatchTarget =
-        prevFocused instanceof HTMLElement && this.element.contains(prevFocused)
-          ? prevFocused
-          : this.element;
-      openPalette = new CommandPalette(this, dispatchTarget);
-    };
-    this.bindAction("open-command-palette", openCommandPalette);
-    // Document-level capture to ensure that the command palette opens even when focus is inside a tool's input element outside viewer.element.
-    this.registerDisposer(
-      registerEventListener(
-        document,
-        "keydown",
-        (event: KeyboardEvent) => {
-          if (event.code === "KeyP" && event.ctrlKey) {
-            event.preventDefault();
-            openCommandPalette();
-          }
-        },
-        { capture: true },
-      ),
-    );
     this.bindAction("deactivate-active-tool", () =>
       this.globalToolBinder.deactivate(),
     );
